refactor!: relocate Clock to tsoracle-server
tsoracle-core::Allocator accepts now_ms: u64 parameters and never
references the Clock trait. The trait + impls existed only for
tsoracle-server's Arc<dyn Clock> builder storage — the
Send + Sync + 'static bound on the trait is required by that
trait-object storage, not by the core algorithm. Workspace-wide, the
only consumers of Clock / SystemClock / MockClock are tsoracle-server
(production + integration tests) and tsoracle-tests (DST suite); the
only reference in tsoracle-consensus was an unrelated doc-comment
match.

- Move Clock + SystemClock + saturating_millis + their unit tests from
  tsoracle-core/src/clock.rs to tsoracle-server/src/clock.rs. Updated
  module doc explains the Send + Sync + 'static bound's true home
  (Arc<dyn Clock> shared across handler tasks).
- Move MockClock + its unit tests to tsoracle-server/src/test_fakes/
  mock_clock.rs, alongside the existing InMemoryDriver / StallableDriver
  / FaultyDriver fakes. MockClock now gates on tsoracle-server's
  existing test-fakes feature instead of inventing a parallel
  test-clock feature.
- Drop the test-clock Cargo feature from tsoracle-core; drop
  features = ["test-clock"] from tsoracle-server's dev-dep on
  tsoracle-core.
- Retarget tsoracle-tests's dst feature: tsoracle-core/test-clock →
  tsoracle-server/test-fakes.
- Rewrite "The Clock contract: advisory only" section in
  tsoracle-core/src/docs/algorithm.md to describe the now_ms: u64
  advisory parameter rather than the (now-relocated) Clock trait. Drop
  the Clock / SystemClock / clock::testing API bullet and the
  test-clock feature row from tsoracle-core's README.

No behavior change. Verification on the full workspace:

- cargo build --workspace --all-features: clean
- cargo test --workspace: all suites pass, 0 failed
- cargo test -p tsoracle-tests --features dst: all turmoil DST suites
  pass, 0 failed
- cargo test -p tsoracle-server --features test-support: clean
- cargo clippy --workspace --all-features -- -D warnings: clean
- cargo doc --workspace --no-deps --all-features: clean for everything
  touched (two pre-existing warnings on main, in unrelated files)
- scripts/check-ts-header.py: 266 files OK

BREAKING CHANGE: removes the public tsoracle_core::Clock,
tsoracle_core::SystemClock, and tsoracle_core::testing re-exports and
the test-clock Cargo feature from tsoracle-core. Downstream consumers
that imported these from tsoracle_core should switch to
tsoracle_server::Clock / tsoracle_server::SystemClock, and
tsoracle_server::test_fakes::MockClock (gated on
tsoracle-server/test-fakes for downstream test code). Custom Clock
impls re-target their `impl Clock for ...` block to the
tsoracle_server::Clock trait — no semantic change to the trait itself.
