+++ crates/tsoracle-core/src/docs/algorithm.md
@@ -18,9 +18,9 @@ Window extension is split into `prepare_window_extension` (sync, no I/O, compute
 
 A new leader at epoch `E_new` must not issue any timestamp at or below any timestamp the prior leader at epoch `E_prev < E_new` could have issued. tsoracle guarantees this through a mandatory failover fence run on every transition into `Leader`: clear serving, drain in-flight extensions, perform a linearized `load_high_water` (the driver MUST guarantee this read reflects all writes durably committed at any prior epoch), compute `serving_floor = max(prior_max + 1, now_ms)`, compute `requested = serving_floor + failover_advance`, durably persist `requested` at the new epoch, seed the allocator with `(serving_floor, actual_high_water, epoch)`, then begin serving. The library refuses `GetTs` with `NOT_LEADER` for the entire duration of the fence. This makes the safety argument direct: the new leader's first issued `physical_ms` is strictly above the prior persisted high-water, and its committed ceiling was durable before serving resumed.
 
-## The Clock contract: advisory only
+## The `now_ms` contract: advisory only
 
-The `Clock` trait's `now_ms` is *advisory*. It tells the allocator how to advance physical time toward the wall clock. The allocator's monotonicity does not depend on clock correctness — a clock jumping backward cannot cause regression because the persisted high-water always wins. A clock pinned far in the past stalls new windows until wall time catches up past the persisted bound. Implementing a custom `Clock` (NTP-disciplined, HLC-extended, externally driven) does not require any changes to the allocator's correctness reasoning.
+Every `Allocator` method that needs physical time accepts it as a `now_ms: u64` parameter. This value is *advisory*: it tells the allocator how to advance physical time toward the wall clock. The allocator's monotonicity does not depend on the value's correctness — a `now_ms` that jumps backward cannot cause regression because the persisted high-water always wins. A `now_ms` pinned far in the past stalls new windows until later samples catch up past the persisted bound. The choice of clock source (`SystemTime`, NTP-disciplined, HLC-extended, externally driven) is therefore the embedder's concern, not the algorithm's, and does not require any changes to the allocator's correctness reasoning.
 
 ## What this is and is not
