+++ crates/tsoracle-server/src/clock.rs
@@ -21,16 +21,16 @@
 //  limitations under the License.
 //
 
-// #[PerformanceCriticalPath]
 //! Source of physical-time milliseconds for the TSO algorithm.
 //!
-//! The Allocator's monotonicity is independent of clock correctness — a clock
+//! The allocator's monotonicity is independent of clock correctness — a clock
 //! jumping backward cannot cause timestamp regression because the persisted
 //! high-water always wins. A clock pinned far in the past stalls new windows
 //! until wall time catches up past the persisted high-water.
-
-#[cfg(any(test, feature = "test-clock"))]
-use core::sync::atomic::{AtomicU64, Ordering};
+//!
+//! The `Send + Sync + 'static` bound exists because the server stores its clock
+//! as an `Arc<dyn Clock>` shared across request-handler tasks; it is not a
+//! property the underlying `tsoracle-core` allocator requires.
 
 pub trait Clock: Send + Sync + 'static {
     /// Milliseconds since Unix epoch.
@@ -69,42 +69,9 @@ fn saturating_millis(d: std::time::Duration) -> u64 {
     u64::try_from(d.as_millis()).unwrap_or(u64::MAX)
 }
 
-#[cfg(any(test, feature = "test-clock"))]
-pub mod testing {
-    use super::*;
-    use std::sync::Arc;
-
-    /// Hand-driven clock for deterministic tests.
-    #[derive(Clone, Default)]
-    pub struct MockClock {
-        ms: Arc<AtomicU64>,
-    }
-
-    impl MockClock {
-        pub fn new(start_ms: u64) -> Self {
-            MockClock {
-                ms: Arc::new(AtomicU64::new(start_ms)),
-            }
-        }
-        pub fn advance(&self, by_ms: u64) {
-            self.ms.fetch_add(by_ms, Ordering::AcqRel);
-        }
-        pub fn set(&self, to_ms: u64) {
-            self.ms.store(to_ms, Ordering::Release);
-        }
-    }
-
-    impl Clock for MockClock {
-        fn now_ms(&self) -> u64 {
-            self.ms.load(Ordering::Acquire)
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use testing::MockClock;
 
     #[test]
     fn system_clock_returns_nonzero() {
@@ -126,25 +93,5 @@ mod tests {
         // bare `as u64` cast would wrap to a small value; saturation must pin
         // it to u64::MAX so a far-future clock never masquerades as the past.
         assert_eq!(saturating_millis(Duration::from_secs(u64::MAX)), u64::MAX);
-    }
-
-    #[test]
-    fn mock_clock_starts_at_seed() {
-        let clock = MockClock::new(42);
-        assert_eq!(clock.now_ms(), 42);
-    }
-
-    #[test]
-    fn mock_clock_advance() {
-        let clock = MockClock::new(100);
-        clock.advance(50);
-        assert_eq!(clock.now_ms(), 150);
-    }
-
-    #[test]
-    fn mock_clock_set() {
-        let clock = MockClock::new(100);
-        clock.set(999);
-        assert_eq!(clock.now_ms(), 999);
     }
 }
