+++ crates/tsoracle-tests/tests/leadership_churn_dst.rs
@@ -45,9 +45,9 @@ use std::time::Duration;
 
 use futures::Stream;
 use tsoracle_consensus::{ConsensusDriver, ConsensusError, LeaderState};
-use tsoracle_core::{Epoch, PeerEndpoint, testing::MockClock};
+use tsoracle_core::{Epoch, PeerEndpoint};
 use tsoracle_proto::v1::GetTsRequest;
-use tsoracle_server::test_fakes::{FaultKind, FaultyDriver, InMemoryDriver};
+use tsoracle_server::test_fakes::{FaultKind, FaultyDriver, InMemoryDriver, MockClock};
 use tsoracle_server::test_support::{wait_until, wait_until_not_serving, wait_until_serving};
 use tsoracle_server::{Server, ServingState};
 use tsoracle_server_testkit::{client, into_sim_parts, serve};
